[Fontations] Enable a build configuration without FreeType dependency
Enable a build option -Dfreetype=disabled to build without FreeType.

This enables removal of the FreeType dependency in downstream projects
such as Chromium. One of -Dfontations=... -Dfreetype=... must be enabled
to configure an indexing backend. If both are enabled, FreeType indexing
is preferred unless the FC_FONTATIONS environment variable is defined.

Set the Mac build to build without FreeType.

In the FreeType-less build disable tests that use .PCF font files. This
format is not supported by the Fontations indexing backend.